Refactor large test-e2e file into smaller components
